fixed tutorial screen
